# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Boards/history": {
      "get": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_GetHistory",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameHistoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Boards/draw": {
      "post": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_Draw",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawNumbersRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Boards/{id}/activate": {
      "patch": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_Activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Boards/{id}/deactivate": {
      "patch": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_Deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Players": {
      "get": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_GetPlayers",
        "parameters": [
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "Sorts",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_CreatePlayer",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}": {
      "get": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_GetPlayerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_UpdatePlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{playerId}/boards": {
      "post": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_CreateBoards",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerBoardsRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playerboard"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}/change-password": {
      "post": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_ChangePassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}/soft-delete": {
      "patch": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_SoftDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}/restore": {
      "patch": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_Restore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GameHistoryDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "boardId": {
            "type": "string"
          },
          "week": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "numbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "participants": {
            "type": "integer",
            "format": "int32"
          },
          "winners": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DrawNumbersRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "numbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "CreatePlayerRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "email",
          "phoneNumber",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 5
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "Playerboard": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "playerid": {
            "type": "string"
          },
          "boardid": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time"
          },
          "iswinner": {
            "type": "boolean"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          },
          "playerboardnumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playerboardnumber"
            }
          }
        }
      },
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "weeknumber": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "startdate": {
            "type": "string",
            "format": "date-time"
          },
          "enddate": {
            "type": "string",
            "format": "date-time"
          },
          "isactive": {
            "type": "boolean"
          },
          "createdat": {
            "type": "string",
            "format": "date-time"
          },
          "drawnnumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Drawnnumber"
            }
          },
          "playerboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playerboard"
            }
          }
        }
      },
      "Drawnnumber": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "boardid": {
            "type": "string"
          },
          "drawnnumber1": {
            "type": "integer",
            "format": "int32"
          },
          "drawnat": {
            "type": "string",
            "format": "date-time"
          },
          "drawnby": {
            "type": "string"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "drawnbyNavigation": {
            "$ref": "#/components/schemas/Admin"
          }
        }
      },
      "Admin": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "passwordhash": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time"
          },
          "isdeleted": {
            "type": "boolean"
          },
          "drawnnumbers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Drawnnumber"
            }
          },
          "processedfundrequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Fundrequest"
            }
          }
        }
      },
      "Fundrequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "playerid": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "transactionnumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time"
          },
          "processedat": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "processedbyadminid": {
            "type": "string",
            "nullable": true
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          },
          "processedbyadmin": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Admin"
              }
            ]
          }
        }
      },
      "Player": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "passwordhash": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time"
          },
          "funds": {
            "type": "number",
            "format": "decimal"
          },
          "isdeleted": {
            "type": "boolean"
          },
          "playerboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playerboard"
            }
          },
          "fundrequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Fundrequest"
            }
          }
        }
      },
      "Playerboardnumber": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "playerboardid": {
            "type": "string"
          },
          "selectednumber": {
            "type": "integer",
            "format": "int32"
          },
          "playerboard": {
            "$ref": "#/components/schemas/Playerboard"
          }
        }
      },
      "CreatePlayerBoardsRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "selectedNumbers"
        ],
        "properties": {
          "selectedNumbers": {
            "type": "array",
            "maxItems": 8,
            "minItems": 5,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "repeatWeeks": {
            "type": "integer",
            "format": "int32",
            "maximum": 52.0,
            "minimum": 1.0
          }
        }
      },
      "UpdatePlayerRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phonenumber": {
            "type": "string",
            "minLength": 3,
            "nullable": true
          }
        }
      },
      "ChangePasswordRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1
          },
          "newPassword": {
            "type": "string",
            "minLength": 6
          }
        }
      }
    }
  }
}